requirements.txt
langchain
langchain_openai
langchain_core
python-dotenv